chore(minor): move huge components into separate files
